[REF] web: process views arch in data manager
A refactoring of the js views is in progress.  Those new views are used
in web studio, and this change is part of it. This is necessary to allow
web studio to correctly fetch data from the server when rendering a
preview.  Note: this is necessary for special widgets (status bar,
selection, radio widgets)