[IMP] ir_module: make icon_image depends on icon path field
For now, the base64 image (binary field) is the
content of the image in /static/src/description.png

The kanban view use 'icon' field, and form view use
'icon_image' field. For module having no icon defined
the default image is not the same on kanban or on
form view.

Moreover, it makes more sence to add this dependency.