[IMP] website(_blog): remove "Style" (blog-style) snippet option
The "bigger text" option was an useless option as redundant with the
summernote editor one.

The "narrow" option always resulted in an ugly layout with bugs in
many contexts.

Note: the styling classes are kept for compatibility for now.

+ Adapt demo data of website_blog which rely on the "narrow" option.
Use up to date snippets with reduced column width to get the same
result.