[IMP] project: Replace 'Cumulative Flow' by a fovorite filter in task report
This report is too advanced and/or complex to be a single menuitem. Furthermore the model was a mess and a horn of plenty for bugs.

The class ProjectTaskHistoryCumulative inherits from the class ProjectTaskHistory, and create a view based on the same model.The orginal class defines a compute field 'end_date' which calls a compute method '_compute_end_date' that will either make a simple assignment or a SQL request if the columns is always opened or folded.

As ProjectTaskHistoryCumulative is declare with '_auto=False' the compute method will try to write on this table, which is obviously wrong because we are trying to write on a view.

This leads to erors when trying to edit a project stage like:

OperationalError: cannot update view "project_task_history_cumulative"
DETAIL:  Views that do not select from a single table or view are not automatically updatable.
HINT:  To enable updating the view, provide an INSTEAD OF UPDATE trigger or an unconditional ON UPDATE DO INSTEAD rule.

or when trying to delete a column which already contains some tasks.

So we removed this these models and view and replaced them by a simple favorite filter in the Tasks Reports