[FIX] hr_timesheet: task creation on project
When creating a task on a project which is accessible to followers only,
an access error pops up.

This comes from the `subtask_project_id` field on the task, which is a
related to a `project.project` field. When a task is created, the ORM
tries to write on the related field, and hence on `project.project`.

opw-687252