[FIX] stock, purchase: fix po confirm of kits BOM
This is an oversight of the revision
cb486d7eb6353a66eea2cb3610141b4f3a778791

In the above revision, which converts the module
`stock` to the new API,
the return value of the method `action_confirm`
has been changed:
 - Before, it returned the list of move ids that have
 been confirmed,
 - After, it returns `True`

The former return value is expected by `mrp`,
when using kits BOM: In this use case,
the purchase of one product can lead to the reception
of multiple products, and therefore the confirmation
of one move can lead to the confirmation of multiple
moves.

In addition, the move that was initially confirmed could
no longer exist after the call of the method `action_confirm`.

The method `action_confirm` of `stock.move`
is expected to return the list of moves that were actually confirmed.

e.g. in purchase, before the conversion to the new API (<= saas-10),
to assign all the moves that were created through
the action confirm of the initial move, and not only the initial moves,
thanks to:
```
move_ids = moves.action_confirm()
moves = self.env['stock.move'].browse(move_ids)
moves.force_assign()
```
The above code has been wrongly converted in the new API with just
```
moves.action_confirm()
moves.force_assign()
```
Which changes the behavior, as, before, all the moves that were
created through `action_confirm` were being assigned, and, after,
only the moves with which the method `action_confirm` has been called
are being assigned.

Besides, with the kits BOM, the move with which `action_confirm` was
being called no longer exists after the call, and attempting to
call the `force_assign` method on it leads to the error
`Record does not exist or has been deleted.`

opw-688231