[IMP] website: make the website dashboard ACL aware
The Website Admin dashboard displays graphs for sales, analytics,etc.
using a (somewhat easily) extensible js framework.

Unfortunately, this framework is ACL unaware, meaning that it will
completely and silently crash if you are missing any of the required
access rights.

To solve this, extensions must now specify not just a graph name
but a necessary group for it and set the value for this group in
the extension of the backend controller.

This commit also
- make the icon visible for the group 'Website Designer and Editor'
  (not juste the Settings group, which is extremely restrictive)
- make the button to set the Google API key only visible to people
  with sufficient access rights