[IMP] product: add sql constraint to uom rounding
rounding field of product_uom now has sql constraint > 0. If set at 0 it rounds
any quantity to 0, which is never a desired result.

Allowing 0 made it so that some users that did not understand the rounding, set
it to 0 and complained that their quantities were messed up.

Also added translations to the error message.