[FIX] website_slides: handle replies to Presentation Published as notes
As the Presentation Published message is posted on a channel every reply
to this message is also posted on the channel. However channels may holds
lot of followers not necessarily interested in being spammed by answers
to that message. We consider now that those replies are notes to avoid
spam. Channel managers can follow the note subtype to be warned about
those messages.