[FIX] website_blog: fix depends for post_date
Before this commit, create a blog post since the backend raise
an error because the post_date is False.

It should never arrive, since it is a compute field that is the
publish date or fallback on the create_date.

But because set post_date, change the published_date, the depend from
publish date cannot be triggered again since it change the post_date.

So _set_post_date don't trigger _compute_post_date. And call it manually
will make 'max recursion depth python'.

The 'hack' to update publish date, without retrigger the _compute_post_date
and set post_date manually, is to use the _write method that don't trigger
all depends fields.

Courtesy of @rco-odoo

This commit closes #13829